pyproject.toml
[tool.poetry]
name = "iris"
version = "0.1.0"
description = ""
authors = ["Niyaz Bureev <bureev.na@phystech.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
scikit-learn = "^1.3.1"
pandas = "^2.1.1"
pre-commit = "^3.4.0"
dvc = {extras = ["gdrive"], version = "^3.26.2"}
hydra-core = "^1.3.2"
onnx = "^1.15.0"
onnxruntime = "^1.16.3"
mlflow = "^2.8.1"
pydantic = "==1.10.13"
scikit-plot = "^0.3.7"
skl2onnx = "^1.15.0"
onnxmltools = "^1.12.0"
tritonclient = "^2.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
